// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/placement.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface Placement {
  id?: string;
  x?: number;
  y?: number;
  z?: number;
  length?: number;
  width?: number;
  height?: number;
  weight?: number;
  stopOrder?: number;
  stackable?: boolean;
  orientation?: number[];
  fragile?: boolean;
  noStack?: boolean;
  bottomOnly?: boolean;
  maxStackWeight?: number;
}